+++ package.json
@@ -18,11 +18,12 @@
     "test:test262": "node run_test262.js",
     "lint": "eslint -c .eslintrc.json ."
   },
-  "dependencies": {
-    "acorn": "^5.2.1"
+  "peerDependencies": {
+    "acorn": "^6.0.0"
   },
   "version": "0.1.1",
   "devDependencies": {
+    "acorn": "^6.0.0",
     "eslint": "^4.13.1",
     "eslint-plugin-node": "^5.2.1",
     "mocha": "^4.0.1",
